% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dx_clone_tree}
\alias{dx_clone_tree}
\title{Clone the directory structure from a DNAnexus folder}
\usage{
dx_clone_tree(remote_path, local_path, remote_subdir, silent)
}
\arguments{
\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}

\item{local_path}{location on the local machine to download the file or folder.}

\item{remote_subdir}{subdirectory relative to the remote_path (see Details)}

\item{silent}{logical; if TRUE does not print message when creating each
folder}
}
\description{
Clone the directory structure from a DNAnexus folder
}
\details{
Recursively traverses the remote folder location given by \code{remote_path} to
recreate the directory tree at the given \code{local_path}. The \code{remote_subdir}
argument is used by the recursive calls to the \code{dx_clone_tree} function and
should be omitted when called by other functions.

The directory tree is recreated at the destination \code{local_path} following
unix conventions: if the \code{remote_path} is a folder ending in "/" the contents
of the folder are downloaded, otherwise the folder itself is downloaded at
the destination \code{local_path}.
}
