% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dx_download}
\alias{dx_download}
\title{Download a file or folder from a DNA nexus project}
\usage{
dx_download(
  remote_path,
  local_path,
  exists = "skip",
  missing = "error",
  incomplete = "wait"
)
}
\arguments{
\item{remote_path}{DNA nexus file identifier or DNA nexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}

\item{local_path}{location on the local machine to download the file or folder.}

\item{exists}{action to take if the file(s) being downloaded already exists
on the local machine. One of "error", "overwrite", or "skip" (default).}

\item{missing}{action to take if the 'remote_path' is not found on DNA nexus.
One of "error" (default), "skip", or "wait".}

\item{incomplete}{action to take if the 'remote_path' points to an incomplete
file on DNA nexus not created by the current DNA nexus job. One of
"error", "skip", or "wait" (default).}
}
\value{
The path on the local machine where the file or files were downloaded
to, replacing the equivalent path given in the 'remote_path' argument.
}
\description{
Extends the \verb{dx download} command line tool with additional functionality
that enables it to optionally wait for incomplete or missing files to be
uploaded by other processes before initiating the download. For programmatic
convenience, the location of the download on the local machine is returned.
}
\details{
When downloading folders, the directory tree is recreated at the destination
'local_path' following unix conventions: if the 'remote_path' is a folder
ending in "/" the contents of the folder are downloaded, otherwise the folder
itself is downloaded at the destination 'local_path'.
}
