% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata.R
\name{dx_get_current_project}
\alias{dx_get_current_project}
\title{Get the ID of the currently selected DNA nexus project}
\usage{
dx_get_current_project()
}
\value{
A string corresponding to a DNA nexus project identifier.
}
\description{
Get the ID of the currently selected DNA nexus project
}
