% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{dx_get_project}
\alias{dx_get_project}
\title{Get the unique identifier of a DNA nexus project}
\usage{
dx_get_project(remote_path)
}
\arguments{
\item{remote_path}{path of an object or folder on DNA nexus project storage.
Syntax: "projectID:/folder/path", "/folder/path", or "folder/path". Paths
are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
A string corresponding to a DNA nexus project identifier.
}
\description{
Get the unique identifier of a DNA nexus project
}
\details{
There are three ways the project identifier is determined:
\enumerate{
\item If no \code{remote_path} is provided, then the ID of the current project is
returned.
\item If \code{remote_path} is provided, and includes a project identifier as part
of the path, returns the project identifier embedded in path.
\item If \code{remote_path} is provided without a project identifier, the ID of the
current project is returned
}
}
