% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata.R
\name{dx_get_project_id}
\alias{dx_get_project_id}
\title{Get the unique ID of a DNA nexus project}
\usage{
dx_get_project_id(metadata)
}
\arguments{
\item{metadata}{project metadata extracted by the \code{dx_get_project_metadata}
function}
}
\value{
an DNA nexus project ID
}
\description{
Get the unique ID of a DNA nexus project
}
