% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_metadata.R
\name{dx_is_container_job}
\alias{dx_is_container_job}
\title{Is the current process running within a containerized DNAnexus job?}
\usage{
dx_is_container_job()
}
\value{
TRUE or FALSE
}
\description{
Is the current process running within a containerized DNAnexus job?
}
\details{
Some DNAnexus jobs, e.g. cloud workstations, are run within a DNAnexus
container. This has two consequences:

(1) When given a path without the project name or project ID, DNAnexus
command line utilities look for the path in the container, instead of
the project storage associated with the job
(2) The current process only has access to the project storage associated
with the job, not the full set of projects the user has permissions to.
}
