% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_metadata.R
\name{dx_is_job}
\alias{dx_is_job}
\title{Is the current process running as part of a DNAnexus job?}
\usage{
dx_is_job()
}
\value{
TRUE or FALSE
}
\description{
Is the current process running as part of a DNAnexus job?
}
