% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{dx_is_project_id}
\alias{dx_is_project_id}
\title{Check whether a string matches the format of a DNAnexus project ID}
\usage{
dx_is_project_id(string)
}
\arguments{
\item{string}{a character vector of length 1}
}
\value{
TRUE or FALSE
}
\description{
See \url{https://documentation.dnanexus.com/developer/api/entity-ids}
}
