% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_paths.R
\name{dx_normalize_path}
\alias{dx_normalize_path}
\title{Get the standardized absolute format of a remote path on DNA nexus}
\usage{
dx_normalize_path(remote_path, return_as_parts = FALSE)
}
\arguments{
\item{remote_path}{a DNA nexus file/folder path with syntax
"projectID:/folder/path", "/folder/path", or "folder/path".}

\item{return_as_parts}{logical; if TRUE, returns the result as a named list
containing the project ID, folder, and name.}
}
\value{
a DNA nexus path with syntax 'project-ID:/path/to/file.ext'
}
\description{
Get the standardized absolute format of a remote path on DNA nexus
}
\details{
This is designed as a low-level helper function - i.e. it avoids making
expensive DNA nexus API calls to 'dx describe' to work out what is at the
remote_path.

If the remote_path is a DNA nexus object identifier rather than a path,
you need to use the \code{dx_path_from_metadata} function after extracting the
object metadata using the \code{dx_get_metadata} function.
}
