% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_paths.R
\name{dx_path_contains_project}
\alias{dx_path_contains_project}
\title{Does the user-provided remote path contain the project name or ID?}
\usage{
dx_path_contains_project(remote_path)
}
\arguments{
\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
TRUE or FALSE
}
\description{
Returns TRUE if the remote path contains a colon (":") and that colon is
not escaped ("\:") (which is the case for files or folders on DNAnexus that
contain a colon).
}
