% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{dx_path_from_metadata}
\alias{dx_path_from_metadata}
\title{Get the path of a file (or other data object) on DNAnexus from metadata}
\usage{
dx_path_from_metadata(metadata)
}
\arguments{
\item{metadata}{file metadata extracted by the internal \code{dx_get_metadata}
function}
}
\value{
a DNAnexus path with syntax 'project-ID:/path/to/file.ext'
}
\description{
Get the path of a file (or other data object) on DNAnexus from metadata
}
