% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{dx_project_protected}
\alias{dx_project_protected}
\title{Determine whether a DNA nexus project has its Protected attribute set to true}
\usage{
dx_project_protected(remote_path)
}
\arguments{
\item{remote_path}{path of an object or folder on DNA nexus project storage.
Syntax: "projectID:/folder/path", "/folder/path", or "folder/path". Paths
are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
Logical; either TRUE or FALSE.
}
\description{
Determine whether a DNA nexus project has its Protected attribute set to true
}
\details{
DNA nexus projects may be configured so that object deletion is restricted to
administrators of the project.
}
