% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{dx_state}
\alias{dx_state}
\title{Get the state of a file (or other data object) on DNA nexus}
\usage{
dx_state(metadata)
}
\arguments{
\item{metadata}{file metadata extracted by the \code{dx_get_metadata} function}
}
\value{
"none" if the file does not exist (or has been deleted; see Details),
"folder" if the location points to a folder, or one of "closed",
"closing", or "open" (see Details).
}
\description{
Get the state of a file (or other data object) on DNA nexus
}
