% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{dx_type}
\alias{dx_type}
\title{Determine the type of entity at a location on a DNAnexus project from metadata}
\usage{
dx_type(metadata)
}
\arguments{
\item{metadata}{file metadata extracted by the internal \code{dx_get_metadata}
function}
}
\value{
"none" if nothing exists at the 'remote_path' provided, "folder" if
the 'remote_path' points to a folder, or the Class attribute (e.g. "file")
if the 'remote_path' points to a DNAnexus object.
}
\description{
Determine the type of entity at a location on a DNAnexus project from metadata
}
