% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{dx_upload}
\alias{dx_upload}
\title{Upload a file or folder to a DNAnexus project}
\usage{
dx_upload(local_path, remote_path = ".", exists = "replace", silent = FALSE)
}
\arguments{
\item{local_path}{path of the file or folder on the local machine to upload.}

\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}

\item{exists}{action to take if the file being uploaded already exists on
DNAnexus. One of "error", "replace" (default), or "skip".}

\item{silent}{logical; if TRUE does not print message on completion of each
file upload.}
}
\description{
Extends the \verb{dx upload} command line tool to enable recursive upload of
folders and resumption of interrupted uploads. Unlike \verb{dx upload}, assumes
the user does not want to create multiple versions of the same file when
uploading a file with the same name that already exists at the 'remote_path'.
}
\details{
When uploading folders, the directory tree is recreated at the destination
'remote_path' location following unix conventions: if the 'local_path' is a
folder ending in "/" the contents of the folder are uploaded otherwise the
folder itself is uploaded at the destination 'remote_path'.
}
\section{File states on DNAnexus}{
Files on DNAnexus can be in three possible states: "closed", "closing", or
"open". Files in the "closed" state are those that are complete and not being
actively written to by \verb{dx upload}. Files in the "closing" state are those
where \verb{dx upload} has finished, and DNAnexus is now finalizing the file.
Files in the "open" state are those that are being actively written to by a
\verb{dx upload} process, or where a \verb{dx upload} process was killed part way
through an upload, e.g. when a low-priority job is interrupted and restarted
by AWS.

To help distinguish between these scenarios and to facilitate checkpointing,
the \code{\link[=dx_upload]{dx_upload()}} function attaches the job ID to the each file as a property
named 'uploaded_by' if run from a DNAnexus job (e.g. from a remote cloud
workstation) so that restarted jobs can identify files which were interrupted
mid-upload and remove them when next queried.
}

