% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata.R
\name{dx_user_can_rm}
\alias{dx_user_can_rm}
\title{Checks whether the user can delete files on a DNA nexus project}
\usage{
dx_user_can_rm(metadata)
}
\arguments{
\item{metadata}{project metadata extracted by the \code{dx_get_project_metadata}
function}
}
\value{
Logical; either TRUE or FALSE.
}
\description{
Checks whether the user can delete files on a DNA nexus project
}
\details{
Returns TRUE if the user has at least "CONTRIBUTE" permissions, or if the
project has the Protected attribute set to true, if the user has "ADMINISTER"
permissions.
}
