% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{throw_file_not_exists_error}
\alias{throw_file_not_exists_error}
\title{Throws an error about the requested object not being found}
\usage{
throw_file_not_exists_error(remote_path)
}
\arguments{
\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
always throws an error
}
\description{
Error message dependent on how the user provided the remote_path and whether
the process is running inside a DNAnexus container job.
}
