% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{assert_dx_exists}
\alias{assert_dx_exists}
\title{Throw an error if the requested object or folder does not exist on DNA nexus}
\usage{
assert_dx_exists(remote_path)
}
\arguments{
\item{remote_path}{path of an object or folder on DNA nexus project storage.
Syntax: "projectID:/folder/path", "/folder/path", or "folder/path". Paths
are relative to the current working directory in the current project
unless otherwise specified.}
}
\description{
Throw an error if the requested object or folder does not exist on DNA nexus
}
