% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{assert_dx_exists}
\alias{assert_dx_exists}
\title{Throw an error if the requested object or folder does not exist on DNAnexus}
\usage{
assert_dx_exists(remote_path, incomplete = TRUE)
}
\arguments{
\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}

\item{incomplete}{logical; if the 'remote_path' points to a file that is
still being written to by another DNAnexus job (see Details), should we
consider this file to exist on DNAnexus?}
}
\description{
Throw an error if the requested object or folder does not exist on DNAnexus
}
