% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata.R
\name{assert_dx_project_permissions}
\alias{assert_dx_project_permissions}
\title{Error if the user does not has sufficient permissions for a given operation}
\usage{
assert_dx_project_permissions(metadata, minimum_permissions)
}
\arguments{
\item{metadata}{project metadata extracted by the \code{dx_get_project_metadata}
function}

\item{minimum_permissions}{minimum permissions user must have on the DNAnexus
project. Must be one of "VIEW", "CONTRIBUTE" or "ADMINISITER"}
}
\description{
Error if the user does not has sufficient permissions for a given operation
}
