% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{dx_exists}
\alias{dx_exists}
\title{Check whether an object or folder exists on a DNA nexus project}
\usage{
dx_exists(remote_path)
}
\arguments{
\item{remote_path}{path of an object or folder on DNA nexus project storage.
Syntax: "projectID:/folder/path", "/folder/path", or "folder/path". Paths
are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
Logical; TRUE or FALSE.
}
\description{
Check whether an object or folder exists on a DNA nexus project
}
