% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{dx_get_metadata}
\alias{dx_get_metadata}
\title{Get the metadata associated with a location on DNAnexus}
\usage{
dx_get_metadata(remote_path)
}
\arguments{
\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
a named list containing information about the target location on
DNAnexus. If it resolves to a DNAnexus object, then the output of
\verb{dx describe --json}. Otherwise a named list containing basic information:
the project ID, parent folder, target file/folder name, and either "folder"
or "none" as the class depending on whether or the remote_path resolves
to a folder on a DNAnexus project or the remote_path does not exist (yet).
}
\description{
If the location resolves to a file that is in the "open" state, removes that
file if it has an "uploaded_by" property matching the current DNAnexus job
ID (see Details).
}
