% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata.R
\name{dx_get_project_metadata}
\alias{dx_get_project_metadata}
\title{Get the metadata associated with a DNA nexus project}
\usage{
dx_get_project_metadata(remote_path)
}
\arguments{
\item{remote_path}{DNA nexus file identifier or DNA nexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
a named list containing the project metadata obtained from the
output of \verb{dx describe --json}
}
\description{
Get the metadata associated with a DNA nexus project
}
