% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata.R
\name{dx_get_project_permissions}
\alias{dx_get_project_permissions}
\title{Determine access permissions of a DNA nexus project}
\usage{
dx_get_project_permissions(remote_path)
}
\arguments{
\item{remote_path}{DNA nexus file identifier or DNA nexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
"NONE", "VIEW", "CONTRIBUTE", or "ADMINISTER", as an ordered factor.
}
\description{
Determine access permissions of a DNA nexus project
}
