% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_metadata.R
\name{dx_get_project_permissions}
\alias{dx_get_project_permissions}
\title{Determine access permissions of a DNA nexus project}
\usage{
dx_get_project_permissions(metadata)
}
\arguments{
\item{metadata}{project metadata extracted by the \code{dx_get_project_metadata}
function}
}
\value{
"NONE", "VIEW", "CONTRIBUTE", or "ADMINISTER", as an ordered factor.
}
\description{
Determine access permissions of a DNA nexus project
}
