% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_path.R
\name{dx_normalize_path}
\alias{dx_normalize_path}
\title{Normalize a DNA nexus path or file ID to an absolute path}
\usage{
dx_normalize_path(remote_path)
}
\arguments{
\item{remote_path}{DNA nexus file identifier or DNA nexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
an absolute remote path with syntax "projectID:/folder/file.txt"
}
\description{
Normalize a DNA nexus path or file ID to an absolute path
}
