% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_paths.R
\name{dx_normalize_path}
\alias{dx_normalize_path}
\title{Get the standardized absolute format of a remote path on DNAnexus}
\usage{
dx_normalize_path(remote_path, return_as_parts = FALSE)
}
\arguments{
\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}

\item{return_as_parts}{logical; if TRUE, returns the result as a named list
containing the project ID, folder, and name.}
}
\value{
a DNAnexus path with syntax 'project-ID:/path/to/file.ext'
}
\description{
Get the standardized absolute format of a remote path on DNAnexus
}
