% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{dx_rm}
\alias{dx_rm}
\title{Delete an object or folder on a DNAnexus project}
\usage{
dx_rm(remote_path, not_exists = "ignore")
}
\arguments{
\item{remote_path}{DNAnexus file identifier or DNAnexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}

\item{not_exists}{action to take if the location being removed does not
exist, either "ignore" or "error".}
}
\description{
Deletes a file or recursively deletes a folder at the given path on DNAnexus.
For projects where the user does not have file deletion ermissions, moves the
file or folder to the trash/ folder at the root folder of the project,
creating that folder if necessary.
}
\details{
Where the user has the necessary permissions to delete files, runs \verb{dx rm -rfa}
on the provided 'remote_path'.

If the user does not have delete permissions and the remote path is a
DNAnexus object (e.g. a file), runs \verb{dx mv} to relocate the object to the
trash/ folder at the project root.

If the user does not have delete permissions and the remote path is a folder,
renames the folder to include the current date and time, then moves that
folder to trash/ with \verb{dx mv}. The renaming step ensures that the folder name
is unique, which necessary to prevent errors when moving to the trash/ folder
as there cannot be multiple folders with the same name in one location on a
DNAnexus project, and \verb{dx mv} does not support merging two folders with the
same name.
}
