% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{dx_type}
\alias{dx_type}
\title{Determine the type of entity at a location on a DNA nexus project}
\usage{
dx_type(remote_path)
}
\arguments{
\item{remote_path}{path of an object or folder on DNA nexus project storage.
Syntax: "projectID:/folder/path", "/folder/path", or "folder/path". Paths
are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
"none" if nothing exists at the 'remote_path' provided, "folder" if
the 'remote_path' points to a folder, or the Class attribute (e.g. "file")
if the 'remote_path' points to a DNA nexus object.
}
\description{
Determine the type of entity at a location on a DNA nexus project
}
