% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{dx_type}
\alias{dx_type}
\title{Determine the type of entity at a location on a DNA nexus project}
\usage{
dx_type(remote_path, from_metadata = FALSE)
}
\arguments{
\item{remote_path}{DNA nexus file identifier or DNA nexus file/folder path
(Syntax: "projectID:/folder/path", "/folder/path", or "folder/path").
Paths are relative to the current working directory in the current project
unless otherwise specified.}

\item{from_metadata}{logical; if TRUE, assumes the \code{remote_path} actually
contains the metadata associated with that location extracted by
\code{\link[=dx_get_metadata]{dx_get_metadata()}}}
}
\value{
"none" if nothing exists at the 'remote_path' provided, "folder" if
the 'remote_path' points to a folder, or the Class attribute (e.g. "file")
if the 'remote_path' points to a DNA nexus object.
}
\description{
Determine the type of entity at a location on a DNA nexus project
}
