% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{dx_user_can_rm}
\alias{dx_user_can_rm}
\title{Checks whether the user can delete files on a DNA nexus project}
\usage{
dx_user_can_rm(remote_path)
}
\arguments{
\item{remote_path}{path of an object or folder on DNA nexus project storage.
Syntax: "projectID:/folder/path", "/folder/path", or "folder/path". Paths
are relative to the current working directory in the current project
unless otherwise specified.}
}
\value{
Logical; either TRUE or FALSE.
}
\description{
Checks whether the user can delete files on a DNA nexus project
}
\details{
Returns TRUE if the user has at least "CONTRIBUTE" permissions, or if the
project has the Protected attribute set to true, if the user has "ADMINISTER"
permissions.
}
